+++ CHANGES.md
@@ -5,6 +5,7 @@
 - Add _request_options to operation call parameters
 - Add informational header to generated files
 - Define `bravado-types` CLI entry point
+- Use enum class for return_types config parameter
 
 ## 0.1.0